backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Datenbank
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Task Queue
redis==5.0.1
celery==5.3.6

# LLM & AI
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.13

# Speech-to-Text (Whisper)
openai-whisper==20231117
# faster-whisper benötigt PyAV was Build-Probleme verursacht
# openai-whisper ist ausreichend und funktioniert ohne PyAV

# Integrationen
requests==2.31.0
httpx==0.25.2
paho-mqtt==1.6.1

# Kamera / Computer Vision
opencv-python-headless==4.9.0.80
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
